Add nlsyscall files.

+++ nlunix/syscall_freebsd.go
@@ -3,7 +3,7 @@ package nlunix
 type SockaddrNetlink struct {
 	Family uint16
 	Pad    uint16
-	Pid    uint16
+	Pid    uint32
 	Groups uint32
 	raw    RawSockaddrNetlink
 }
